Also copy shim/grub to rootfs from fallback if its alpine
When we dont find a grub/shim in the rootfs we go for the fallback one
which is the one provided by the grub2 luet packages.

But for the install we wont have any artifacts in the rootfs to use for
alpine.

So in the case of alpine copy those fallback artifacts into the rootfs
so we have something to work on

Signed-off-by: Itxaka <itxaka@kairos.io>