Fix ubuntu grub.efi naming
Looks like even if the original name of the ubuntu artifacts are
grubx64.efi, the shim still looks for grub.efi to boot from.

This patch copies the grub artifact with the original name AND also with
the default grub.efi as a fallback in case the shim looks for that name
instead.

Signed-off-by: Itxaka <itxaka@kairos.io>