Move build-uki test to a new e2e suite
to avoid stubbing everything. enki is a cli and has behavior we want to
test that depends on the environment it runs in. E.g.

when specific binaries are not found in PATH, enki will return an error.

In order to test such cases, we need to be able to control the env in
which enki is run. The simplest way is to run inside a container and for
that we need to first build an enki binary and then run it inside a
specified container.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>