Add a --artifact flag to generate different output artifacts
adds a --artifact falg that allows to set the output type.
Currently iso, artifact and container are supported and will generate
the proper artifacts.

 - iso: Will generate an iso in the output dir
 - artifact: Will generate the UKI EFI files, configs and keys in the
   output dir
 - container: Will generate a container image witht he artifacts inside
   them (Valid for upgrade)

This changes the --output flag from a specific iso file to a dir, to be
able to generate the output artifacts using that dir as a base. Name is
now generated automatically based on the version obtained from the
source.

Iso will now be kairos_VERSION.iso
Artifact/config names are kept the same as before
Container will be kairos_uki:$VERSION

Signed-off-by: Itxaka <itxaka@kairos.io>