Use a middle dir to store needed artifacts to ukify
we were mangling the dir where we extracted the source with the
artifacts needed to ukify the image, so we ended up with artifacts int
eh final uki image, like the vmlinuz.

This just uses a middle temp dir to store the kernel and initramfs
outside of the rootfs so there is no possibility of polluting it.

Saves about 15Mb

Signed-off-by: Itxaka <itxaka@kairos.io>