Ensure valid and non-problematic efi filenames for UKIs
When building a UKI, the cmdline becomes part of the filename.
This causes errors when slashes (/) are present, as they are illegal on
Linux filesystems.
This happens for example when extending the cmdline with a config_url.
example: `--extend-cmdline="config_url=http://host:port"`.

Other characters may not be forbidden, but might cause problems if they
are interpreted by a shell or are just inconvenient.
To be safe, I changed the implementation from only replacing spaces with
underscores to only allowing alphanumeric, dot, underscore and hyphen
characters and replace all others with an underscore.

Signed-off-by: ZauberNerd <zaubernerd@zaubernerd.de>