Allow customizing artifact name for build-uki command
The `build-iso` command accepts a `--name` flag to specify the basename
of the generated iso file, while the `build-uki` command uses the value
of `KAIROS_RELEASE` from the `/etc/os-release` file.
This commit introduces a `--name` flag for the `build-uki` command to
allow specifying the basename of the generated .iso or .tar file.
When the output type is `uki` the `--name` flag is ignored, because the
uki output type write the uki filesystem into the output folder.

Signed-off-by: ZauberNerd <zaubernerd@zaubernerd.de>