Bundle bios boot artifacts
This patch makes enki bundle the needed bios boot artifacts for livecd
This uses embed to embed them in the final binary and not have to depend
on anything external anymore.

This includes the core.img (which is grub+cdboot) and the boot_hybrid
image. Plus the default livecd grub config file

This should result in enki being to be run as a simple binary without
needing anything in the system (other than squashfs and xorriso to build
the iso)

No artifacts from efi nor grub ones shold be needed anymore

Signed-off-by: Itxaka <itxaka@kairos.io>