Refactor enki build-uki command
so that it can be used as a library from Auroraboot
as part of https://github.com/kairos-io/kairos/issues/1633

The following anti-patterns make it very hard to re-use the enki
functions:

- viper is used as a global store, randomly accessing everything from anywhere
- we pass around too many "configs": the BuildConfig, the embedded
  kairos-agent config, the BuildUKIAction (which is essentially yet another config),

It's nearly impossible to construct the proper input for any function
because we don't know what part of the configs available, is actually needed
and used.

This is a first refactoring that should at least let the caller use the
BuildUKIAction without needing to populate a global "viper" object.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>